requirements.txt
accelerate==0.24.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioprometheus==23.12.0
aiosignal==1.3.1
altair==5.4.1
antlr4-python3-runtime==4.9.3
anyio==4.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arxiv==2.1.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
Brotli==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cmake==3.30.3
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
deepdiff==8.0.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
duckduckgo-search==3.9.9
effdet==0.4.1
einops==0.7.0
emoji==2.12.1
environs==9.5.0
et-xmlfile==1.1.0
faiss-cpu==1.7.4
fastapi==0.109.0
feedparser==6.0.10
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fschat==0.2.35
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.25.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==7.2.1
iniconfig==2.0.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langchain==0.0.354
langchain-community==0.0.20
langchain-core==0.1.23
langchain-experimental==0.0.47
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.0.87
language_data==1.2.0
latex2mathml==3.77.0
layoutparser==0.3.4
lit==18.1.8
llama-index==0.9.35
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.0
markdownify==0.11.6
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
metaphor-python==0.1.23
minio==7.2.8
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.8.1
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.18
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
olefile==0.47
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
openai==1.9.0
opencv-python==4.10.0.84
openpyxl==3.1.5
orderly-set==5.2.2
orjson==3.10.7
packaging==23.2
pandas==2.0.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pgvector==0.3.3
pi_heif==0.16.0
pikepdf==8.4.1
Pillow==9.5.0
pluggy==1.5.0
portalocker==2.10.1
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.13
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pymilvus==2.3.6
PyMuPDF==1.23.16
PyMuPDFb==1.23.9
pypandoc==1.13
pyparsing==3.1.4
pypdf==5.0.0
pypdfium2==4.30.0
pytest==7.4.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
quantile-python==1.1
rapidfuzz==3.9.7
rapidocr-onnxruntime==1.3.8
ray==2.36.0
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
sgmllib3k==1.0.0
shapely==2.0.6
shortuuid==1.0.13
simplejson==3.19.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.4.8
sse-starlette==1.8.2
starlette==0.35.0
streamlit==1.30.0
streamlit-aggrid==0.3.4.post3
streamlit-antd-components==0.3.1
streamlit-chatbox==1.1.11
streamlit-feedback==0.1.3
streamlit-option-menu==0.3.12
streamlit_modal==0.1.0
strsimpy==0.2.1
svgwrite==1.4.3
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.5.2
timm==1.0.9
tokenizers==0.15.2
toml==0.10.2
torch==2.0.0
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4.1
tqdm==4.66.1
transformers==4.37.2
transformers-stream-generator==0.0.4
triton==2.0.0
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
unstructured==0.15.12
unstructured-client==0.22.0
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
validators==0.34.0
vllm==0.2.7
wasabi==1.1.3
watchdog==3.0.0
watchfiles==0.24.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.3.4
websockets==12.0
Werkzeug==3.0.4
wrapt==1.16.0
xformers==0.0.23.post1
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.11.1
youtube-search==2.1.2
zipp==3.20.2
